api/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pymupdf
python-dotenv
langchain-core
langchain-text-splitters
langchain-community
langchain-openai
pydantic
supabase
httpx
PyJWT
boto3
mangum
requests
